refactor : User, UserSummary 더미 데이터 제거
