--- src/main/java/com/fourthread/ozang/module/domain/feed/dto/dummy/User.java
@@ -1,9 +1,0 @@
-package com.fourthread.ozang.module.domain.feed.dto.dummy;
-
-import com.fourthread.ozang.module.domain.BaseEntity;
-import jakarta.persistence.Entity;
-
-@Entity
-public class User extends BaseEntity {
-
-}
