--- src/main/java/com/fourthread/ozang/module/domain/feed/dto/dummy/UserSummary.java
@@ -1,5 +1,0 @@
-package com.fourthread.ozang.module.domain.feed.dto.dummy;
-
-public class UserSummary {
-
-}
