+++ src/main/java/com/fourthread/ozang/module/domain/feed/entity/Feed.java
@@ -1,8 +1,8 @@
 package com.fourthread.ozang.module.domain.feed.entity;
 
 import com.fourthread.ozang.module.domain.BaseUpdatableEntity;
-import com.fourthread.ozang.module.domain.feed.dto.dummy.User;
 import com.fourthread.ozang.module.domain.feed.dto.dummy.Weather;
+import com.fourthread.ozang.module.domain.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ManyToOne;
