+++ src/main/java/com/fourthread/ozang/module/domain/feed/entity/FeedComment.java
@@ -1,9 +1,9 @@
 package com.fourthread.ozang.module.domain.feed.entity;
 
+import com.fourthread.ozang.module.domain.user.entity.User;
 import static org.hibernate.annotations.OnDeleteAction.CASCADE;
 
 import com.fourthread.ozang.module.domain.BaseUpdatableEntity;
-import com.fourthread.ozang.module.domain.feed.dto.dummy.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ManyToOne;
