+++ src/main/java/com/fourthread/ozang/module/domain/feed/entity/FeedLike.java
@@ -1,7 +1,7 @@
 package com.fourthread.ozang.module.domain.feed.entity;
 
 import com.fourthread.ozang.module.domain.BaseEntity;
-import com.fourthread.ozang.module.domain.feed.dto.dummy.User;
+import com.fourthread.ozang.module.domain.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ManyToOne;
